Desktop/uruntanima/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
ultralytics==8.3.161
pandas==2.1.3
pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3 